% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_neighbour_rank_lists.R
\name{build_neighbour_rank_lists}
\alias{build_neighbour_rank_lists}
\title{Create lists of neighbours for each genetic marker on the basis of ranks}
\usage{
build_neighbour_rank_lists(query_cluster, rank_df)
}
\arguments{
\item{rank_df}{}
}
\description{
Create lists of neighbours for each genetic marker on the basis of ranks
}
