% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate_synteny_blocks.R
\name{cross_validate_synteny_blocks}
\alias{cross_validate_synteny_blocks}
\title{Perform cross validation of synteny blocks via subsampling}
\usage{
cross_validate_synteny_blocks(map_list, max_cluster_range,
  max_nn_dist = 15, min_block_size = 3, plots = FALSE,
  buffer_cM = 1, n_sims = 100)
}
\arguments{
\item{n_sims}{}
}
\description{
Perform cross validation of synteny blocks via subsampling
}
