% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_rate_1.R
\name{estimate_error_rate_1}
\alias{estimate_error_rate_1}
\title{Estimate mapping error rate from empirical data}
\usage{
estimate_error_rate_1(map_list, linear_regions)
}
\arguments{
\item{linear_regions}{}
}
\description{
Estimate mapping error rate from empirical data
}
