% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_rate_2.R
\name{estimate_error_rate_2}
\alias{estimate_error_rate_2}
\title{Estimate mapping noise rate from empirical data}
\usage{
estimate_error_rate_2(map_list, colinear_chrs)
}
\arguments{
\item{colinear_chrs}{}
}
\description{
Estimate mapping noise rate from empirical data
}
