% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_clusters.R
\name{group_clusters}
\alias{group_clusters}
\title{Interavely consolidate and group marker cluster neighbourhoods}
\usage{
group_clusters(neighbour_list)
}
\arguments{
\item{neighbour_list}{list containing the neighbours of each marker cluster, e.g. generated by build_neighbour_rank_lists}
}
\value{
a list containing group assignments for each marker cluster
}
\description{
Interavely consolidate and group marker cluster neighbourhoods
}
