% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_map.R
\name{make_sim_map}
\alias{make_sim_map}
\title{Simulate two genetic maps that differ at a single inversion}
\usage{
make_sim_map(n_positions, n_markers, inversion_size, error_rate_1,
  error_rate_2, bend_factor = 1)
}
\arguments{
\item{bend_factor}{}
}
\description{
Simulate two genetic maps that differ at a single inversion
}
