% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_stats.R
\name{plot_summary_stats}
\alias{plot_summary_stats}
\title{Plot goodness-of-fit metrics for multiple parameter combination}
\usage{
plot_summary_stats(summary_stats, stats_to_plot)
}
\arguments{
\item{stats_to_plot}{}
}
\description{
Plot goodness-of-fit metrics for multiple parameter combination
}
