% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntR-package.R
\docType{package}
\name{syntR-package}
\alias{syntR}
\alias{syntR-package}
\title{syntR: Detection of synteny blocks via comparison of genetics maps}
\description{
Detection of synteny blocks via comparison of genetics maps
}
\author{
\strong{Maintainer}: Katherine Ostevik \email{kate.ostevik@gmail.com}

Authors:
\itemize{
  \item Kieran Samuk \email{ksamuk@gmail.com}
}

}
\keyword{internal}
