% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters.R
\name{test_parameters}
\alias{test_parameters}
\title{Test the performance of a range of tuning parameters}
\usage{
test_parameters(map_list, max_cluster_range_list, max_nn_dist_list,
  min_block_size = 2)
}
\arguments{
\item{min_block_size}{}
}
\description{
Test the performance of a range of tuning parameters
}
