% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters_sim_data.R
\name{test_parameters_sim_data}
\alias{test_parameters_sim_data}
\title{Test the performance of tuning parameters on simulated data}
\usage{
test_parameters_sim_data(max_cluster_range_list, max_nn_dist = 15,
  min_block_size = 2, plots = FALSE, n_sims, buffer_cM, n_positions,
  n_markers, inversion_size, error_rate_1, error_rate_2)
}
\arguments{
\item{error_rate_2}{}
}
\description{
Test the performance of tuning parameters on simulated data
}
