% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_blocks.R
\name{adjust_blocks}
\alias{adjust_blocks}
\title{Adjust synteny blocks to prevent spanning chromosomes and overlaps}
\usage{
adjust_blocks(map_list, mark_df, clust_df)
}
\arguments{
\item{map_list}{a map list object}

\item{mark_df}{a marker dataframe}

\item{clust_df}{a cluster dataframe}
}
\value{
an adjust map list object
}
\description{
Adjust synteny blocks to prevent spanning chromosomes and overlaps
}
