% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_initial_blocks.R
\name{find_initial_blocks}
\alias{find_initial_blocks}
\title{Perform initial clustering of genetic markers}
\usage{
find_initial_blocks(clust_df)
}
\arguments{
\item{clust_df}{}
}
\description{
Perform initial clustering of genetic markers
}
