% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_clusters.R
\name{group_clusters}
\alias{group_clusters}
\title{Perform cluster re-grouping via the friends-of-friends approach}
\usage{
group_clusters(neighbour_list)
}
\arguments{
\item{neighbour_list}{}
}
\description{
Perform cluster re-grouping via the friends-of-friends approach
}
