% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_one_map.R
\name{make_one_map}
\alias{make_one_map}
\title{Collate two genetic maps into a single dataframe}
\usage{
make_one_map(map, space_size = 20, order_chr1 = NULL,
  order_chr2 = NULL, max_length1 = NULL, max_length2 = NULL,
  flip_chrs = NULL)
}
\arguments{
\item{flip_chrs}{}
}
\description{
Collate two genetic maps into a single dataframe
}
