% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_one_map.R
\name{make_one_map}
\alias{make_one_map}
\title{Title}
\usage{
make_one_map(map, space_size = 20, map1_chr_order = NULL,
  map2_chr_order = NULL, map1_max_chr_lengths = NULL,
  map2_max_chr_lengths = NULL, flip_chrs = NULL)
}
\arguments{
\item{flip_chrs}{}
}
\description{
Title
}
