% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_maps}
\alias{plot_maps}
\title{Plot a two-dimension representation of two genetic maps}
\usage{
plot_maps(map_df, sp1_chrom_breaks, sp2_chrom_breaks, pch_val = 16,
  cex_val = 0.75, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot a two-dimension representation of two genetic maps
}
