% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_stats.R
\name{plot_summary_stats}
\alias{plot_summary_stats}
\title{Create a heat map of summary statistics generated by test_parameters}
\usage{
plot_summary_stats(summary_stats, stats_to_plot)
}
\arguments{
\item{summary_stats}{list, containing the output of test_parameters}

\item{stats_to_plot}{character, the names of the stastistics to plot}
}
\value{
a heatmap of the clustering performance for the selected statistics
}
\description{
Create a heat map of summary statistics generated by test_parameters
}
