% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_blocks.R
\name{split_blocks}
\alias{split_blocks}
\title{Split synteny blocks that span chromosomal boundaries}
\usage{
split_blocks(block_target, cent_df, map_list)
}
\arguments{
\item{map_list}{}
}
\description{
Split synteny blocks that span chromosomal boundaries
}
