% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters.R
\name{test_parameters}
\alias{test_parameters}
\title{Test the performance of a range of clustering parameters for a given map_list}
\usage{
test_parameters(map_list, max_cluster_range_list, max_nn_dist_list,
  min_block_size = 2, progress_bar = FALSE)
}
\arguments{
\item{map_list}{list, a map_list object generated by find_synteny_blocks}

\item{max_cluster_range_list}{list of length 2, the minimum and maxmium values for max_clust_range}

\item{max_nn_dist_list}{list of length 2, the minimum and maxmium values for max_nn_dist}

\item{min_block_size}{numeric of length 1, the minimum size of a synteny block}
}
\value{
a list containing a summary of clustering performance for each parameter combination
}
\description{
Test the performance of a range of clustering parameters for a given map_list
}
\examples{

}
